Relocate Feather metadata helper into plugin tree
